flow : rebuild main + added new files + basic listener
