+++ src/flow/basicUpdatesListener.go
@@ -1,0 +1,5 @@
+package main
+
+type basicUpdatesListener interface {
+	handleUpdates()
+}
